backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2
httpx==0.28.1
email-validator==2.2.0
